requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.1
redis==4.6.0
Pillow==10.0.1
django-htmx==1.17.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-decouple==3.8
django-extensions==3.2.3
whitenoise==6.5.0
gunicorn==21.2.0
django-cors-headers==4.3.0
django-filter==23.2
moviepy==1.0.3
ffmpeg-python==0.2.0
python-magic==0.4.27
django-storages==1.14
boto3==1.28.57
django-cleanup==8.0.0
django-mptt==0.14.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
asgiref==3.7.2